+++ src/main/php/lang/types/Float.class.php
@@ -1,7 +1,5 @@
 <?php namespace lang\types;
 
-
-
 /**
  * The Float class wraps a value of the type float
  */